backend/foodgram_project/requirements.txt
abstract==2021.11.29.1
asgiref==3.5.0
author==0.0.0.0.dev20200926
backports.zoneinfo==0.2.1
base32hex==1.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
colouration==2020.7.26
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
filelock==3.4.2
graphviz==0.19.1
gunicorn==20.1.0
idna==3.3
ingredients==0.3.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pip-upgrade-tool==0.7.2
platformdirs==2.4.1
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
slug==2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
value==0.1.0
virtualenv==20.13.0
webcolors==1.11.1
